==> scripts/prod-testing/2_mad_prod_rfishprod.R <==
# Productivity calcs - rfishprod version (Morais & Bellwood 2020)

source("scripts/prod-testing/1_mad_data_prep.R")

ls()
head(mada.prod)


# Check "size"
summary(mada.prod$size)
#min = 1cm, max = 100cm, median = 12cm

plot(mada.prod$size)
# Seems to be quite a lot of small fish. Few > 40cm.



# Check if any individuals are larger than species max size:
length(mada.prod$size)
# 5219

summary(mada.prod[mada.prod$size > mada.prod$MaxSizeTL,])
# 57
#57/5219   
# 1.0 %

# or equal to max size
summary(mada.prod[mada.prod$size == mada.prod$MaxSizeTL,])
# 147
#147/5219
# 2.8 %

#quite a lot. These need to be reduced to equal MaxSizeTL (prod= exactly 0) 
#  OR 0.1cm below MaxSizeTL (tiny prod values)

mada.prod$size2 <- ifelse(mada.prod$size >= mada.prod$MaxSizeTL, mada.prod$MaxSizeTL-0.1, mada.prod$size)
head(mada.prod)
#Now use size2 for further analyses



###########################################################################



# Use rfishprod package to estimate productivity (i.e. biomass production over time)

#library(devtools)
#devtools::install_github("renatoamorais/rfishprod")

library(rfishprod)

mada.prod <- tidytrait (mada.prod, sey.trait2) 

# from Morais & Bellwood 2018:
fmod <- formula (~ sstmean + MaxSizeTL + Position + Diet + Method) 



# Predicting Kmax, the standardised VBGF parameter (Recommendation: use 100s to 1000s iterations) 
# (takes a while)
# see ?predKmax
datagr2 <- predKmax (mada.prod, 
                    dataset = db,
                    fmod = fmod,
                    niter = 10,
                    return = 'pred')

datagr2 <- datagr2$pred



# Positioning fish in their growth trajectory 
# i.e. what's the size they're supposed to have on the next day? 
datagr2$L.1day <- with (datagr2, applyVBGF (Lmeas = size2,
                                          Lmax = MaxSizeTL,
                                          Kmax = Kmax))

head(datagr)
#each fish has grown a tiny amount (in length).


#Calculate age estimates:
datagr$EstAge <- (1/datagr$Kmax)*log((datagr$MaxSizeTL)/((1-datagr$size2/datagr$MaxSizeTL)*datagr$MaxSizeTL))
summary(datagr$EstAge)
#(if lengths not reduced below MaxSizeTL, will have infinite age values)


#Calculate growth with von Bertalanffy growth function (VBGF), using ages:
#VBGF:  Lt = Lmax*(1-exp(-K*t))
# t = Estimated age + time interval
# Morais & Bellwood (2019), interval = 1day (i.e. 1/365)

# Calculate over a full year: age + (1:365)/365

#Age to add for each day of the year:
age <- (1:365)/365

#Create table for new length: 1 column per day, 1 row per individual fish
VB_lngth <-  matrix(ncol=length(age), nrow=nrow(datagr), dimnames=list(NULL, paste("Day", 1:365, sep="_")))

# for each individual, calculate new length for each day using VBGF formula
for(u in 1:nrow(datagr)) {
  VB_lngth[u, ] <- datagr$MaxSizeTL[u]*(1-exp(-datagr$Kmax[u]*(datagr$EstAge[u] + age)))
}

#Now have a matrix of lengths for each day of the year per fish
range(VB_lngth)
#  1.009271   101.314415


#Convert lengths to weights using a & b coefficents:
VB_wt <-  apply(VB_lngth, 2, function(x) datagr$a*(x^datagr$b))  
head(VB_wt)
nrow(VB_wt)
# 5219 rows
# units = mass in grams (per fish)

biomass.g <- (datagr$a * datagr$size2^datagr$b)
VB_wt2 <- cbind(biomass.g, VB_wt)
head(VB_wt2)
colnames(VB_wt2)[1] <- "Day_0"
dim(VB_wt2)
# 5219   366
# First column = mass when observed during survey, then after 1 day, 2 days etc.

#create new object:
VB_prod_wt <- VB_wt2

#calculate daily mass produced (productivity)
for(u in 2:ncol(VB_prod_wt))
  for(v in 1:nrow(VB_prod_wt)) {
    VB_prod_wt[v,u] <- VB_wt2[v,u] - VB_wt2[v,(u-1)]
  }

#Delete first column (still total mass in g)
VB_prod_wt2 <- VB_prod_wt[,2:ncol(VB_prod_wt)]
head(VB_prod_wt2) 
# table of daily production in g/individual over a year (not cumulative)


# Now sum row totals to get POTENTIAL productivity over 1 year (i.e. without accounting for mortality):
prod_yr <- rowSums(VB_prod_wt2)
prod_yr  # 1 value per fish (row)
summary(prod_yr)
#   Min.   1st Qu.    Median      Mean   3rd Qu.       Max. 
# 0.1418   11.5866   26.8228   60.8488   77.3024  1171.2566 


#Take value after 1 day as daily productivity value:
prod_day <- VB_prod_wt2[,1] 
prod_day
summary(prod_day)
#      Min.   1st Qu.    Median      Mean   3rd Qu.      Max. 
# 0.0002209 0.0286485 0.0635776 0.1480279 0.1862414 3.1542922  

prod.per.fish <- cbind(prod_day,prod_yr)
head(prod.per.fish, 10)

# Now have daily and annual potential productivity estimates per fish.

#write.csv(prod.per.fish, "data/Potential-prod-test_rfishprod.csv", row.names=F)

#prod.per.fish <- read.csv("data/Potential-prod-test_rfishprod.csv")

mada.prod2 <- cbind(mada.prod, prod.per.fish, Kmax=datagr$Kmax)
# WCS Madagascar data, with potential prod columns added in
# NOTE: must join with mada.prod, not original raw data where individuals are in different order.

write.csv(mada.prod2, "data/wcs/madagascar_potential-prod-test_rfishprod.csv", row.names=F)
